requirements.txt
sentence-transformers==2.4.0
langchain
streamlit
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_community
langchain-openai==0.3.9
langchain_experimental
requests
-e .
